extract LoginUseCaseOutputComposer to production code
