+++ test/core/navigator_factory/login_composer/login_use_case_output_composer_test.dart
@@ -1,29 +1,11 @@
 import 'package:flutter_test/flutter_test.dart';
+import 'package:thepos/core/login_composer/login_use_case_output_composer.dart';
 import 'package:thepos/features/login/data/login_service/login_errors.dart';
 import 'package:thepos/features/login/data/login_use_case/login_use_case_output.dart';
 import 'package:thepos/features/login/data/models/login_result.dart';
 
 import '../../../login/data/helpers/shared_test_helper.dart';
 import '../../../login/presentation/helpers/shared_helpers.dart';
-
-class LoginUseCaseOutputComposer extends LoginUseCaseOutput {
-  LoginUseCaseOutputComposer(this.outputs);
-
-  final List<LoginUseCaseOutput> outputs;
-  @override
-  void onLoginSuccess(LoginResult result) {
-    for (final LoginUseCaseOutput output in outputs) {
-      output.onLoginSuccess(result);
-    }
-  }
-
-  @override
-  void onLoginFail(LoginErrors error) {
-    for (final LoginUseCaseOutput output in outputs) {
-      output.onLoginFail(error);
-    }
-  }
-}
 
 void main() {
   test('compose multiple outputs should delegate success message', () {
